requirements.txt
torch
torchaudio
librosa
numpy
scikit-learn
pandas
scipy
praat-parselmouth
transformers
datasets
accelerate
evaluate
soundfile
tqdm
pyyaml
matplotlib
joblib
git+https://github.com/MorenoLaQuatra/yaml_config_override
gdown
